feat: add address search API endpoint
Added a new API endpoint /address_search/ which can be used to search
addresses with given text and language provided as query params. The
result is a GeoJSON containing matching addresses and their coordinates.

Under the hood the search uses the Digitransit API address search. In
the past our UI executed address search requests against the Digitransit
API directly, but that is not possible anymore as Digitransit started
requiring usage of an API key.