fix: make address importing work in OpenShift
The municipality importer from Munigeo, which is needed to import
addresses, writes temporary data to /app/data directory. That didn't
work in OpenShift because it runs containers with an arbitrary user who
doesn't have permission to write to that directory. The recommended way
to handle that situation is to allow the root group to write to that
directory, and that is exactly what was done.