feat: bring back permission to modify events for contractors
Related to ticket PS-159 contractors' permission to modify their events
was removed. However, it turned out that contractors still need to at
least be able to view all the info on the UI's modify event page, and
the easiest fix for that is to bring back the whole event modifying
ability.

Refs: PS-184